fix: move feedback to about
